Refactored the escaping stuff into A cpp file and added several comments.
